Add files via upload
This is the whole epic of downloading a backup from loging in till actually downloading and displaying the path to the file. This file can also be used as a building block for a bigger epic (like download backup and restore in PgAdmin) that is why i put it in the mainfunctions directory.