Update uploadpackage.js
There was a bug in the setversionNumber function. It is not possible to check if a JsHandle returned from a page.EvaluateHandle exists. So it fails to check if the versionnumber is empty. 

I now first added a page.Evaluate to check if the input is empty and if it exists, it calls a page.EvaluateHandle to get the actual input Handle to be able to type 1 in it.